// src/redux/productPage/types.ts
export enum ActionType {
    GET_ALL_PRODUCT = 'GET_ALL_PRODUCT',
    PRODUCTS_SPINNER = 'PRODUCTS_SPINNER',
    GET_PRODUCT_DESCRIPTION = 'GET_PRODUCT_DESCRIPTION',
    GET_RELATED_PRODUCTS = 'GET_RELATED_PRODUCTS',
    GET_FEATURED_PRODUCT = 'GET_FEATURED_PRODUCT',
    GET_LATEST_PRODUCT = 'GET_LATEST_PRODUCT',
    GET_EXCLUSIVE_PRODUCT = 'GET_EXCLUSIVE_PRODUCT',
    QUOTES = 'QUOTES',
    GET_SEARCH_PRODUCT = 'GET_SEARCH_PRODUCT',
}